Updated the cast insertion algorithm.
It now applies unbox in more places.

I realized why the following rule will not work:

    G |-t2 : A2                     G |- B1 ~ B2
    G |- t1 : C  fun(C) = A1 -> B1  G |- A2 ~ A1
    -------------------------------------------- :: app
    G |- t1 t2 : B2

Notice that this rule casts the type B1 to B2, but during
type *construction* we don't know the type B2, and hence, we do not have
two types to cast between like we do with A1 and A2.

Thus, this rule is not decidable.