Adding a separate MFPS draft.
The MFPS page limit is only 15 pages with a very large margin.  So we
just do not have the room to explain all that we have done.

So here is the plan.  We will submit a short version of our paper on
just the model and simply typed grady to MFPS.  Then we will take the
MFPS draft and extend it with the gradual type system we have developed
and submit that as an extended version to JFP which has no page limit.