Getting closer to the full first draft.
Had an interesting realization.

We had box and unbox as explicit features of the surface language.
These are used to cast the result of a function application, because the
application rule does not implicitly cast the result.

I realized that these can infact be *defined* in the Surface language
and so they do need to be added explicitly to the language.  The
following defines:

box A t = ((\x:?.x) t)
unbox A t = ((\x:A.x) t)

squash S t = ((\x:?.t) t)
split S t = ((\x:S.t) t)

The pattern is, use the identity function to force the function
application rule to cast the argument.  Then after cast insertion we
obtain the respective cast.  That is,

(box A t) = ((\x:?.x) t) => ((\x:?.x) (box A t')) ~> box A t'

Here I am denoting cast insertion by "=>", and so the program on the
right of "=>" is the Core Grady program where "t => t'".

Thus, in gradual typing the surface language *does* support explicit
casts by default.