Restated each result just above their proof in the appendix.
This makes it easier for readers without having to flip the page so
much.

Used this new LaTex package called thmtools and thm-restate.  This
allowed me to declare a lemma or theorem as "restatable" which then
allows one to restate the lemma or theorem with the same number.  Also,
it keeps the full lemma definition in one place, and then, assigns a
command to the entire lemma for restating it.  This means that if we
need to change the result we only need to change it in one place. :-)